Refactored displayDueDate and updateNextDueDate. Removed react-day-picker from backend. Remove personal developer comments.
