+++ package.json
@@ -25,8 +25,7 @@
     "passport": "^0.5.0",
     "passport-auth0": "^1.4.1",
     "pg": "^8.7.1",
-    "prettier": "^2.4.1",
-    "react-day-picker": "^7.4.10"
+    "prettier": "^2.4.1"
   },
   "scripts": {
     "start": "nodemon server/server.js",
